# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BillPro - Store Billing & Reports API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/weekly:
    get:
      operationId: getWeeklyReport
      tags: [reports]
      summary: Get weekly progress report
      parameters:
        - name: weekOffset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Weekly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReport"
  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly sales report data
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
...
components:
  schemas:
...
    LineItem:
      type: object
      required: [id, type, name, quantity, unitPrice, total]
      properties:
        id:           { type: integer }
        type:         { type: string }
        productId:    { type: ["integer", "null"] }
        serviceId:    { type: ["integer", "null"] }
        name:         { type: string }
        description:  { type: ["string", "null"] }
        quantity:     { type: number }
        unitPrice:    { type: number }
        discount:     { type: number }
        total:        { type: number }
...
    Sale:
      type: object
      required: [id, invoiceNumber, status, subtotal, tax, total, createdAt]
      properties:
        id:             { type: integer }
        invoiceNumber:  { type: string }
        customerId:     { type: ["integer", "null"] }
        customerName:   { type: ["string", "null"] }
        customerEmail:  { type: ["string", "null"] }
        status:         { type: string }
        subtotal:       { type: number }
        taxRate:        { type: number }
        tax:            { type: number }
        discount:       { type: number }
        total:          { type: number }
        notes:          { type: ["string", "null"] }
        paymentMethod:  { type: ["string", "null"] }
        dueDate:        { type: ["string", "null"] }
        paidAt:         { type: ["string", "null"] }
        items:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        createdAt:      { type: string }
...
    DailyRevenue:
      type: object
      required: [date, revenue, salesCount]
      properties:
        date:       { type: string }
        revenue:    { type: number }
        salesCount: { type: integer }
    WeeklyReport:
      type: object
      required: [weekStart, weekEnd, totalRevenue, totalSales, newCustomers, topItems, dailyBreakdown]
      properties:
        weekStart:      { type: string }
        weekEnd:        { type: string }
        totalRevenue:   { type: number }
        totalSales:     { type: integer }
        newCustomers:   { type: integer }
        avgSaleValue:   { type: number }
        topItems:
          type: array
          items:
            $ref: "#/components/schemas/TopItem"
        dailyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DailyRevenue"
    MonthlyReport:
      type: object
      required: [year, month, totalRevenue, totalSales, newCustomers, salesList]
      properties:
        year:           { type: integer }
        month:          { type: integer }
        totalRevenue:   { type: number }
        totalSales:     { type: integer }
        newCustomers:   { type: integer }
        avgSaleValue:   { type: number }
        revenueByWeek:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyRevenueSummary"
        salesList:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
    WeeklyRevenueSummary:
      type: object
      required: [week, revenue, salesCount]
      properties:
        week:       { type: integer }
        revenue:    { type: number }
        salesCount: { type: integer }
    TopItem:
      type: object
      required: [name, type, totalRevenue, totalQty]
      properties:
        name:         { type: string }
        type:         { type: string }
        totalRevenue: { type: number }
        totalQty:     { type: number }
...